Allow first time users to set up as admins
Added simple single use page to add the first user to the database
this user will be assumed to be the admin and the page will not be
accessible after use. Created simple sessions controller to allow
users to log in and out, first steps in allowing admin control and
user sign up.