Created simple supporter level system
Created supporter levels system allowing for supporter levels to be
created, edited and destroyed. Supporter levels are shown thorough
the admin dashboard with no need for a direct link to the show page
at present. Tests created to monitor the use of these throuh the
admin dash in simple terms.