Suggestions now belong to the user that created them
Added user_id and ARR to suggestions so that suggestions now belong
to the user that created them. Additionally the vote count is now
stored on the suggestion for simplicity in sorting and minimising
the number of quereies. On vote create and destroy the count is
increased and decreased respectively with a new suggesiton
defaulting to 0 votes. Finally some simple styling has been added
to the show page.