Added the core admin setup to the app
Admins can now fully control the homepage and the content that is
displayed. DB is pre-populated with opening text helping new users
to get started and then it is over written. Validation has been
applied accross the platform and basic styling has been update.
Additionally the app now uses the default bootstrap styling with a
heavy greyscale over for aesthetic.